requirements.txt
# Production dependencies
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
duckdb==1.1.1
pandas==2.2.3
requests==2.32.3
streamlit==1.39.0
numpy==2.1.1
python-dotenv==1.0.1
boto3==1.34.162
openpyxl
chardet
langgraph
pydantic>=2
sqlglot>=25
dbt-core>=1.7
dbt-duckdb>=1.7
fastapi
uvicorn
pyyaml
